fix: Use creator og image on all sub routes of /creators
