+++ apps/main-landing/src/app/creators/page.tsx
@@ -1,26 +1,6 @@
 import { ScrollArea } from '@idriss-xyz/ui/scroll-area';
-import { Metadata } from 'next';
-
-import { metadata as landingMetadata } from '@/app/layout';
 
 import Content from './content';
-
-// ts-unused-exports:disable-next-line
-export const metadata: Metadata = {
-  ...landingMetadata,
-  description:
-    'Creator monetization app that helps you earn more with instant payouts and near-zero platform cuts.',
-  openGraph: {
-    title: 'Make more, grow faster, take control',
-    description:
-      'Creator monetization app that helps you earn more with instant payouts and near-zero platform cuts.',
-    images: [
-      {
-        url: '/og-creators.png',
-      },
-    ],
-  },
-};
 
 // ts-unused-exports:disable-next-line
 export default function Landing() {
